feat: rename to recycling-locator
