+++ .env.example
@@ -14,8 +14,8 @@ SENTRY_PROJECT=
 # production version
 VITE_PUBLIC_PATH=/
 
-# The base url for the widget API
-VITE_WIDGET_API=http://rl.test/widget/
+# The base url for the API
+VITE_RECYCLING_LOCATOR_API=http://rl.test/widget/
 
 # The public Rest API key for the Here Maps API
 VITE_HERE_MAPS_PLACES_KEY=
