+++ index.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/recycling-locator-widget.css" />
-    <title>Recycling Locator widget</title>
+    <link rel="stylesheet" href="/recycling-locator.css" />
+    <title>Recycling Locator Widget</title>
 
     <style>
       :root {
@@ -24,8 +24,8 @@
   </head>
   <body>
     <main>
-      <h1>Recycling Locator widget</h1>
-      <recycling-locator-widget></recycling-locator-widget>
+      <h1>Recycling Locator Widget</h1>
+      <recycling-locator></recycling-locator>
     </main>
 
     <script type="module" src="/src/index.tsx"></script>
