+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "recycling-locator-widget",
+  "name": "recycling-locator",
   "version": "0.0.0",
   "description": "Find places to recycle, refill or reuse items.",
   "type": "module",
