+++ public/recycling-locator.css
@@ -2,13 +2,13 @@
  * Optional styles registered in the global scope
  * Prevents Flash Of Unstyled Content (FOUC)
  */
-recycling-locator-widget {
+recycling-locator {
   display: block;
   height: 540px;
 }
 
 @media (min-width: 768px) {
-  recycling-locator-widget {
+  recycling-locator {
     height: 640px;
   }
 }
