+++ readme.md
@@ -1,9 +1,9 @@
-# Recycling Locator Widget
+# Recycling Locator
 
 ## Welsh language support
 
 Use the `locale` attribute to set the language. `cy` is currently the only supported language code.
 
 ```html
-<recycling-locator-widget locale="cy"></recycling-locator-widget>
+<recycling-locator locale="cy"></recycling-locator>
 ```
