+++ src/components/composition/Layout/Layout.css
@@ -8,7 +8,7 @@ locator-layout {
   grid-template-rows: auto auto 1fr;
   height: 100%;
 
-  .standalone & {
+  .recycling-locator-variant-standalone & {
     height: 100cqh;
   }
