+++ src/components/content/Logo/Logo.tsx
@@ -1,7 +1,7 @@
 import register from 'preact-custom-element';
 
 import config from '@/config';
-import { CustomElement } from '@/types/custom-element';
+import { CustomElement } from '@/types/customElement';
 
 export default function Logo() {
   return (
