+++ src/styles/tokens/container.css
@@ -10,7 +10,7 @@
 
   /**
    * This matches the height setup in the optional FOUC preventing css file
-   * /public/recycling-locator-widget.css
+   * /public/recycling-locator.css
    */
   --container-height: 540px;
   @media (min-width: 768px) {
