+++ standalone.html
@@ -13,10 +13,10 @@
   </head>
   <body>
     <main>
-      <recycling-locator-widget
+      <recycling-locator
         variant="standalone"
         basename="/standalone"
-      ></recycling-locator-widget>
+      ></recycling-locator>
     </main>
 
     <script type="module" src="/src/index.tsx"></script>
